requirements.txt
langchain
python-dotenv
streamlit
openai
tiktoken
unstructured
googletrans==3.1.0a0
assemblyai